Fix bluetoothd logging.
This was using syslog, which is a broken API in Android.  Using the logcat
API so we can get lower-level BT debugging.  Use killall -USR2 bluetoothd
to enable debug logging.

(Fixing syslog is a pain: syslog() tries to treat /proc/kmsg as a
socket, which doesn't work; it can't write to it anyway due to lack of
permissions needing kernel changes to fix; and it's in Bionic which probably
shouldn't depend on liblog, which is the API it should really be using.
Forget about it and just make apps use liblog.)